Fix errors from Doxygen and builds
re #15518

+++ MantidQt/CustomInterfaces/src/Muon/MuonSequentialFitDialog.cpp
@@ -455,7 +455,7 @@ void MuonSequentialFitDialog::continueFit() {
 
       // If multiple groups/periods, set up simultaneous fit
       if (datasetsPerRun > 1) {
-        for (int i = 1; i < workspacesToFit.size(); i++) {
+        for (size_t i = 1; i < workspacesToFit.size(); i++) {
           std::string suffix = boost::lexical_cast<std::string>(i);
           fit->setPropertyValue("InputWorkspace_" + suffix, workspacesToFit[i]);
           fit->setProperty("WorkspaceIndex_" + suffix, 0);
